theme_importer: Make `syntax` more generic
This allows us to reuse some functionality between importers without carrying the theme system-specific baggage.
